+++ core/src/main/java/org/vertexium/ExtendedDataRow.java
@@ -2,6 +2,16 @@ package org.vertexium;
 
 public interface ExtendedDataRow extends VertexiumObject {
     /**
+     * Meta property name used for sorting and aggregations
+     */
+    String TABLE_NAME = "__extendedDataTableName";
+
+    /**
+     * Meta property name used for sorting
+     */
+    String ROW_ID = "__extendedDataRowId";
+
+    /**
      * Id of the row
      */
     ExtendedDataRowId getId();