+++ core/src/main/java/org/vertexium/ExtendedDataRowBase.java
@@ -1,11 +1,17 @@
 package org.vertexium;
 
+import com.google.common.collect.Lists;
+import org.vertexium.mutation.ElementMutation;
+import org.vertexium.property.MutablePropertyImpl;
 import org.vertexium.util.ConvertingIterable;
 import org.vertexium.util.FilterIterable;
 
 import java.util.Iterator;
 
 public abstract class ExtendedDataRowBase implements ExtendedDataRow {
+    private transient Property rowIdProperty;
+    private transient Property tableNameProperty;
+
     @Override
     public abstract ExtendedDataRowId getId();
 
@@ -39,6 +45,11 @@ public abstract class ExtendedDataRowBase implements ExtendedDataRow {
 
     @Override
     public Property getProperty(String key, String name, Visibility visibility) {
+        if (ExtendedDataRow.ROW_ID.equals(name)) {
+            return getRowIdProperty();
+        } else if (ExtendedDataRow.TABLE_NAME.equals(name)) {
+            return getTableNameProperty();
+        }
         for (Property property : getProperties()) {
             if (isMatch(property, key, name, visibility)) {
                 return property;
@@ -72,6 +83,12 @@ public abstract class ExtendedDataRowBase implements ExtendedDataRow {
 
     @Override
     public Iterable<Property> getProperties(String key, String name) {
+        if (ExtendedDataRow.ROW_ID.equals(name)) {
+            return Lists.newArrayList(getRowIdProperty());
+        } else if (ExtendedDataRow.TABLE_NAME.equals(name)) {
+            return Lists.newArrayList(getTableNameProperty());
+        }
+
         return new FilterIterable<Property>(getProperties()) {
             @Override
             protected boolean isIncluded(Property prop) {
@@ -111,6 +128,12 @@ public abstract class ExtendedDataRowBase implements ExtendedDataRow {
 
     @Override
     public Object getPropertyValue(String key, String name, int index) {
+        if (ExtendedDataRow.ROW_ID.equals(name)) {
+            return getRowIdProperty().getValue();
+        } else if (ExtendedDataRow.TABLE_NAME.equals(name)) {
+            return getTableNameProperty().getValue();
+        }
+
         Iterator<Object> values = getPropertyValues(key, name).iterator();
         while (values.hasNext() && index > 0) {
             values.next();
@@ -128,5 +151,37 @@ public abstract class ExtendedDataRowBase implements ExtendedDataRow {
             return getId().compareTo(((ExtendedDataRow) o).getId());
         }
         throw new ClassCastException("o must be an " + ExtendedDataRow.class.getName());
+    }
+
+    protected Property getRowIdProperty() {
+        if (rowIdProperty == null) {
+            rowIdProperty = new MutablePropertyImpl(
+                    ElementMutation.DEFAULT_KEY,
+                    ExtendedDataRow.ROW_ID,
+                    getId().getRowId(),
+                    null,
+                    null,
+                    null,
+                    null,
+                    FetchHint.ALL
+            );
+        }
+        return rowIdProperty;
+    }
+
+    protected Property getTableNameProperty() {
+        if (tableNameProperty == null) {
+            tableNameProperty = new MutablePropertyImpl(
+                    ElementMutation.DEFAULT_KEY,
+                    ExtendedDataRow.ROW_ID,
+                    getId().getTableName(),
+                    null,
+                    null,
+                    null,
+                    null,
+                    FetchHint.ALL
+            );
+        }
+        return tableNameProperty;
     }
 }